Relocate to features directory and fix paths
