+++ apps/app/src/features/openai/client/components/RagSearchButton.module.scss
@@ -1,7 +1,7 @@
 @use '@growi/core-styles/scss/bootstrap/init' as bs;
 @use '@growi/core-styles/scss/variables/growi-official-colors';
 @use '@growi/ui/scss/atoms/btn-muted';
-@use './button-styles';
+@use '~/client/components/PageControls/button-styles';
 
 .btn-rag-search :global {
   @extend %btn-basis;
