+++ apps/app/src/features/openai/client/components/RagSearchButton.tsx
@@ -2,7 +2,8 @@ import React, { useCallback } from 'react';
 
 import { NotAvailableForGuest } from '~/client/components/NotAvailableForGuest';
 import { useIsAiEnabled } from '~/stores-universal/context';
-import { useRagSearchModal } from '~/stores/rag-search';
+
+import { useRagSearchModal } from '../stores/rag-search';
 
 import styles from './RagSearchButton.module.scss';
