+++ apps/app/src/features/openai/client/stores/rag-search.ts
