+++ apps/app/src/features/openai/interfaces/ai.ts
