+++ apps/app/src/features/openai/server/services/client-delegator/get-client.ts
@@ -1,4 +1,4 @@
-import { OpenaiServiceType } from '~/interfaces/ai';
+import { OpenaiServiceType } from '../../../interfaces/ai';
 
 import { AzureOpenaiClientDelegator } from './azure-openai-client-delegator';
 import type { IOpenaiClientDelegator } from './interfaces';
