+++ apps/app/src/features/openai/server/services/client.ts
