+++ apps/app/src/features/openai/server/services/embeddings.ts
